+++ src/requisition-order-create/messages_en.json
@@ -1,0 +1,8 @@
+{
+  "requisition.orderCreate": "Create Orders",
+  "requisition.orderCreate.submitted": "Orders created successfully",
+  "requisition.orderCreate.createdOrderSent.success": "Offline created order sent successfully.",
+  "requisition.orderCreate.createdOrderSent.error": "Error occurred while sending offline created order.",
+  "requisition.orderCreate.draftUpdate.success": "Draft order update success.",
+  "requisition.orderCreate.draftUpdate.error": "Error occurred while updating draft order."
+}
