+++ src/requisition-order-create/order-create-helper-functions.jsx
@@ -1,6 +1,3 @@
-import TrashButton from '../react-components/buttons/trash-button';
-import InputCell from '../react-components/table/input-cell';
-
 export const pushNewOrder = (fetchedOrder, setOrderParams, stockCardSummaryRepositoryImpl, setOrders) => {
     const orderParams = {
         programId: fetchedOrder.program.id,
@@ -45,42 +42,6 @@ const setOrderWithSoh = (page, fetchedOrder, setOrders) => {
     setOrders(prevState => [...prevState, orderWithSoh]);
 };
 
-export const getTableColumns = () => {
-    return [
-        {
-            Header: 'Product Code',
-            accessor: 'orderable.productCode'
-        },
-        {
-            Header: 'Product',
-            accessor: 'orderable.fullProductName'
-        },
-        {
-            Header: 'SOH',
-            accessor: 'soh',
-            Cell: ({ value }) => (<div className="text-right">{value}</div>)
-        },
-        {
-            Header: 'Quantity',
-            accessor: 'orderedQuantity',
-            Cell: (props) => (
-                <InputCell
-                    {...props}
-                    numeric
-                    key={`row-${_.get(props, ['row', 'original', 'orderable', 'id'])}`}
-                />
-            )
-        },
-        {
-            Header: 'Actions',
-            accessor: 'id',
-            Cell: ({ row: { index }, deleteRow }) => (
-                <TrashButton onClick={() => deleteRow(index)} />
-            )
-        }
-    ]
-};
-
 export const getUpdatedOrder = (selectedOrderable, order) => {
     const newLineItem = {
         orderedQuantity: '',
@@ -104,3 +65,14 @@ export const getUpdatedOrder = (selectedOrderable, order) => {
 
     return updatedOrder;
 };
+
+export const createOrderDisabled = (orders) => {
+    const mappedOrders = orders
+        .map((order) => Boolean(order.orderLineItems && order.orderLineItems.length && order.id !== undefined));
+    return mappedOrders.length === 0 || mappedOrders.includes(false)
+}
+
+export const saveDraftDisabled = (orders) => {
+    const mappedOrders = orders.map((order) => order.id !== undefined);
+    return mappedOrders.length === 0 || mappedOrders.includes(false);
+}
