+++ src/requisition-order-create/order-create.scss
@@ -11,6 +11,15 @@
     }
   }
 
+  .requisition-info {
+    width: calc(100% - 2em);
+    margin: 1em 0 0 1em;
+    position: relative;
+    background-color: $background-color-alt;
+    border: 1px solid $border-color;
+    border-radius: $border-radius;
+  }
+
   .order-create-form {
     .section {
       margin: 0.5em 0;
@@ -65,14 +74,13 @@
     max-height: 40vh;
 
     .search {
-
       .input-wrapper {
-        display:inline-block;
+        display: inline-block;
         position: relative;
 
         &:before {
-          font-family: 'FONTAWESOME';
-          content: '\f0d7'; // fa-caret-down
+          font-family: "FONTAWESOME";
+          content: "\f0d7"; // fa-caret-down
           position: absolute;
           left: 1em;
           color: $light-grey;
@@ -114,7 +122,7 @@
 
     .select {
       position: absolute;
-      z-index:100;
+      z-index: 100;
       border: $light-grey 1px solid;
       box-shadow: 0 0.25em 1em #e4e4e4;
       background-color: #fff;
@@ -153,8 +161,8 @@
             width: 100%;
 
             &.is-selected {
-                background-color: $accent-color;
-                color: #fff;
+              background-color: $accent-color;
+              color: #fff;
             }
           }
         }
